Merapikan pengkodean manajemen supplier.

+++ app/Http/Controllers/Api/ApiCompanyController.php
@@ -29,46 +29,59 @@ class ApiCompanyController extends Controller
         ]);
         return response()->json([
             'status' => 'success',
-            'data' => $data
+            'data'   => $data
         ]);
     }
     public function show($id)
     {
         $data = Company::findOrFail($id);
-        return response()->json([
-            'status' => 'success',
-            'data' => $data
-        ]);
+        return response()->json(['data'   => $data]);
     }
     public function update(Request $request, $id)
     {
         $this->validate($request, [
-            'code' => 'required|unique:companies,code,'.$id,
+            'code' => 'required|unique:companies,code,' . $id,
             'name' => 'required'
         ]);
         $data = Company::findOrFail($id);
-        if (count($request->type) == 2) {
-            $request->type == 'both';
+        if (null !== $request->type) {
+            if (count($request->type) == 2) {
+                $request->type == 'both';
+            }
         }
         $data->update($request->all());
         return response()->json([
             'status' => 'success',
-            'data' => $data
+            'data'   => $data
         ]);
     }
     public function destroy($id)
     {
         $data = Company::findOrFail($id);
-        $data->delete();
-        return response()->json([
-            'status' => 'success',
-            'message' => 'Data berhasil dihapus',
-            'data' => null
-        ]);
+        if ($data->items->count() > 0) {
+            return response()->json([
+                'status'  => 'canceled',
+                'message' => 'Supplier ' . $data->name . ' masih memiliki produk yang terdaftar.',
+                'data'    => null,
+            ]);
+        }
+        else {
+            $data->delete();
+            return response()->json([
+                'status'  => 'success',
+                'message' => 'Supplier ' . $data->name . ' berhasil dihapus.',
+                'data'    => null
+            ]);
+        }
     }
     public function fetchSuppliers()
     {
-        $suppliers = Company::whereIn('type', ['supplier', 'both'])->get();
+        $suppliers = Company::query()->whereIn('type', ['supplier', 'both']);
         return CompanyResource::collection($suppliers);
+    }
+    public function getByType($type)
+    {
+        $data = Company::query()->whereIn('type', [$type, 'both']);
+        return $data;
     }
 }
